+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fixes package name in documentation
 - Fixes Intro code chunk to be up to date
 - Changes snake_case variables to camelCase
+- Add Navbar case for no projectUrl with story documentation
 
 ## v0.7.1